Add Postgres support (DB_DIALECT=postgres) alongside MySQL
Switches the deployment target from Aiven MySQL to Neon Postgres.
Verified end-to-end against a real Postgres 16 container: all 8
services boot, sync their schema, seed defaults, log in, and issue
JWTs; finance-service's document-numbering sequence was additionally
load-tested for uniqueness under 8 concurrent claims.

- shared/src/dialect.js: isMySQL(sequelize) helper.
- services/*/src/config/database.js: dialect and driver now follow
  DB_DIALECT (default mysql, unchanged); skip the MySQL-only
  CREATE-DATABASE bootstrap step on Postgres, where managed providers
  pre-provision a fixed database the app user can't sibling-create.
- services/*/src/index.js: every legacy migration written in raw
  MySQL-only SQL (backtick quoting, AUTO_INCREMENT, CHANGE COLUMN,
  INFORMATION_SCHEMA, etc.) now skips itself on Postgres. These exist
  only to evolve an existing MySQL install forward; sequelize.sync()
  already creates the correct schema from the current models on a
  fresh database of any dialect, so there is nothing for them to do.
  ORM-only seed steps (seedRolesAndPermissions, admin bootstrap,
  seedNotificationConfigs, seedUnitCatalog) are untouched and still
  run on both dialects.
- services/finance-service/src/utils/documentSequence.js: the one
  piece of MySQL-only SQL that runs on every request rather than at
  boot (atomic document-number claims via LAST_INSERT_ID) now has a
  real Postgres path using INSERT ... ON CONFLICT ... RETURNING.
- services/crm-service/src/models/lead.js: DataTypes.TINYINT.UNSIGNED
  -> SMALLINT.UNSIGNED. Postgres has no native TINYINT and Sequelize
  can't translate it there (found by actually booting against
  Postgres — crm-service failed with `type "tinyint" does not exist`).
- services/auth-service/src/index.js: the Google-OAuth DB-settings
  lookup now quotes identifiers per-dialect via quoteIdentifier
  instead of hardcoded backticks.
- package.json: add pg + pg-hstore alongside the existing mysql2.
- render.yaml, DEPLOY_RENDER.md: retargeted at Neon (auto-resumes on
  the next query after its 5-minute idle suspend, vs Supabase's
  week-long pause that needs a manual dashboard restore — a better
  fit for a demo that may sit untouched between client calls.
- cred.env.example: documents DB_DIALECT and the Postgres port.

Known, unfixed risk: several services sync with { alter: true }.
Sequelize's ENUM handling differs between MySQL and Postgres when
altering an EXISTING enum column's values later — not a concern for
a fresh deploy, worth revisiting before treating Postgres as anything
more than the current fresh-install target.

EOF
)

+++ services/crm-service/src/models/lead.js
@@ -24,8 +24,11 @@ module.exports = (sequelize, DataTypes) => {
     purchase_window: { type: DataTypes.STRING },
     lead_thermal: { type: DataTypes.STRING, defaultValue: 'Cold' },
 // Add ai_score field
+    // SMALLINT rather than TINYINT: Postgres has no native TINYINT, and
+    // Sequelize cannot translate it there (unlike UNSIGNED, which it can
+    // just drop) — the score only needs to hold 0-100 either way.
     ai_score: {
-      type: DataTypes.TINYINT.UNSIGNED,
+      type: DataTypes.SMALLINT.UNSIGNED,
       defaultValue: 0,
     },
